修改导航页存储位置增加{{paginator.prefix_url}}和{{paginator.page_url}}两个变量

+++ lib/function.php
@@ -3,8 +3,10 @@ function mkfolder($path)	{
 	return is_writeable($path) || mkdir($path, 0777, true);
 }
 
-function paginator($post) {
+function paginator($post,$type,$name) {
 	global $site;
+	$prefix_url = $site['url'];
+	$prefix_url .= !empty($type) ? '/'.$type.'/'.$name.'/' : '/';
 	$perpage = $site['config']['posts_per_page'];
 	$total_pages = ceil(count($post)/$perpage);
 	for($i=0;$i<$total_pages;$i++) {
@@ -13,16 +15,18 @@ function paginator($post) {
 		if($i<1) $has_previous = false; 
 		$pagpost = array_slice($post, $i*$perpage, $perpage);
 		$paginator[$i+1] =  array(
+			'prefix_url' => $prefix_url,
 			'per_page' => $perpage,
 			'object_list'=>$pagpost, 
 			'total_pages'=> $total_pages, 
 			'page' => $i+1, 
+			'page_url' => $prefix_url.'page/'.($i+1),
 			'previous_page' => $i,
 			'pre_page' => $i, 
-			'previous_page_url' => $i.'.html',
-			'pre_page_url' => $i.'.html', 
+			'previous_page_url' => $prefix_url.'page/'.$i,
+			'pre_page_url' => $prefix_url.'page/'.$i, 
 			'next_page' => $i+2, 
-			'next_page_url' => ($i+2).'.html', 
+			'next_page_url' => $prefix_url.'page/'.($i+2), 
 			'has_next' => $has_next, 
 			'has_previous' => $has_previous,
 			'has_pre' => $has_previous
@@ -32,12 +36,10 @@ function paginator($post) {
 }
 
 
-function removeDir($dirName) { 
-        foreach(glob($dirName.'/*') as $dir) {
-                foreach(glob($dir.'/*') as $f) @unlink($f);
-                rmdir($dir);
-        }
+function DDir($dirName) {
+	foreach(glob($dirName.'/*') as $dir)	is_dir($dir) ? DDir($dir) && rmdir($dir) : unlink($dir);
 }
+
 
 function daily($dir) {
 	$doc = array();
